packages/camphr_transformers/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "camphr_transformers"
version = "0.1.3"
description = ""
authors = [ "Yohei Tamura <tamuhey@gmail.com>",]

[tool.versionup]
files = [ "camphr_transformers/VERSION.py",]
commit = true
tag = true
tag_prefix = "camphr_transformers/"

[tool.poetry.dependencies]
python = "^3.7"
transformers = "^4.8"
camphr = "^0.10.0"
pytextspan = "^0.5.4"

[tool.poetry.extras]
torch = [ "torch",]

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^21.7b0"
hypothesis = "^6.14.5"
isort = "^5.9.3"

[tool.poetry.dependencies.torch]
version = "^1.8.0"
optional = true
